encapsulate js
